feat: 상품 재고 조회 API
내부 통신을 위한 상품 재고 조회 API 생성

+++ src/main/java/com/example/stockservice/domain/dto/StockRes.java
@@ -6,5 +6,5 @@ import lombok.Setter;
 
 @AllArgsConstructor
 public class StockRes {
-    private Long stock;
+    private int stock;
 }
